+++ Kernel/ESTD/Algorithm.h
@@ -17,19 +17,19 @@ namespace ESTD
 	}
 
 	template<typename T>
-	T AlignUp(const T& data, uint32_t alignment)
+	T AlignUp(const T& data, size_t alignment)
 	{
 		return ((data % alignment) ? ((data + alignment - 1) & ~(alignment - 1)) : data);
 	}
 
 	template<typename T>
-	T AlignDown(const T& data, uint32_t alignment)
+	T AlignDown(const T& data, size_t alignment)
 	{
 		return ((data % alignment) ? (data & ~(alignment - 1)) : data);
 	}
 
 	template<typename T>
-	bool IsAligned(const T& data, uint32_t alignment)
+	bool IsAligned(const T& data, size_t alignment)
 	{
 		return (data % alignment == 0);
 	}
