+++ Kernel/Memory/PhysicalAddress.h
@@ -4,5 +4,5 @@
 
 namespace Kernel
 {
-	using PhysicalAddress = uint32_t;
+	using PhysicalAddress = uintptr_t;
 }
