+++ Kernel/Memory/VirtualMemoryAllocator.h
@@ -16,24 +16,24 @@ namespace Kernel
 		class VirtualMemoryAllocator
 		{
 		public:
-			bool Init(VirtualAddress startAddress, uint32_t size);
+			bool Init(VirtualAddress startAddress, size_t size);
 			VirtualAddress AllocatePage();
-			VirtualAddress AllocatePages(uint32_t pageCount);
+			VirtualAddress AllocatePages(size_t pageCount);
 			bool FreePage(VirtualAddress address);
-			bool FreePages(VirtualAddress address, uint32_t pageCount);
+			bool FreePages(VirtualAddress address, size_t pageCount);
 			void ReservePage(VirtualAddress address);
-			void ReservePages(VirtualAddress address, uint32_t pageCount);
-			uint32_t GetPageSize() const;
+			void ReservePages(VirtualAddress address, size_t pageCount);
+			size_t GetPageSize() const;
 			
 		private:
-			void SetRegionState(VirtualAddress address, uint32_t sizeBytes, VirtualPageState state);
-			uint32_t AddressToPageNumber(VirtualAddress address) const;
-			VirtualAddress PageNumberToAddress(uint32_t pageNumber) const;
+			void SetRegionState(VirtualAddress address, size_t sizeBytes, VirtualPageState state);
+			size_t AddressToPageNumber(VirtualAddress address) const;
+			VirtualAddress PageNumberToAddress(size_t pageNumber) const;
 
 		private:
-			uint32_t m_PageSize;
-			uint32_t m_MemorySizeBytes;
-			uint32_t m_PageCount;
+			size_t m_PageSize;
+			size_t m_MemorySizeBytes;
+			size_t m_PageCount;
 			VirtualAddress m_StartAddress;
 			VirtualAddress m_EndAddress;
 			ESTD::Bitmap<131072> m_Bitmap;
